+++ packages/models/src/updater.spec.ts
@@ -1,3 +1,5 @@
+import type { IOmnichannelRoom } from '@rocket.chat/core-typings';
+
 import { UpdaterImpl } from './updater';
 
 test('updater typings', () => {
@@ -14,6 +16,20 @@ test('updater typings', () => {
 		};
 		e: string[];
 	}>({} as any);
+
+	const omnichannel = new UpdaterImpl<IOmnichannelRoom>({} as any);
+	omnichannel.addToSet('v.activity', 'asd');
+	// @ts-expect-error
+	omnichannel.addToSet('v.activity', 1);
+	// @ts-expect-error
+	omnichannel.addToSet('v.activity', {
+		asdas: 1,
+	});
+
+	// @ts-expect-error
+	omnichannel.addToSet('v.activity.asd', {
+		asdas: 1,
+	});
 
 	updater.addToSet('e', 'a');
 
@@ -160,7 +176,7 @@ test('it should add items to array', async () => {
 		{
 			_id: 'test',
 		},
-		{ $addToSet: { $each: { a: ['b', 'c'] } } },
+		{ $addToSet: { a: { $each: ['b', 'c'] } } },
 	);
 });