+++ ee/packages/ddp-client/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @rocket.chat/ddp-client
 
+## 0.3.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies [264d7d5496, b8e5887fb9]:</summary>
+
+  - @rocket.chat/rest-typings@6.11.0
+  - @rocket.chat/api-client@0.2.3
+  </details>
+
+## 0.3.3-rc.6
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/rest-typings@6.11.0-rc.6
+  - @rocket.chat/api-client@0.2.3-rc.6
+  </details>
+
+## 0.3.3-rc.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/rest-typings@6.11.0-rc.5
+  - @rocket.chat/api-client@0.2.3-rc.5
+  </details>
+
+## 0.3.3-rc.4
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/rest-typings@6.11.0-rc.4
+  - @rocket.chat/api-client@0.2.3-rc.4
+  </details>
+
+## 0.3.3-rc.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/rest-typings@6.11.0-rc.3
+  - @rocket.chat/api-client@0.2.3-rc.3
+  </details>
+
 ## 0.3.3-rc.2
 
 ### Patch Changes